# docs/admin-api.openapi.yaml
...
openapi: 3.1.0
info:
  title: Mockxy Admin API
  version: 1.2.0
  summary: Control surface of a running Mockxy mock server — catalog, response variants, collections, data files, traffic monitor and runtime switches.
  description: |
    Every action of the Mockxy UI is a call to this API: there is nothing the interface can do
    that a script or an agent cannot. Typical uses are seeding a workspace, resetting state
    between e2e tests, importing an updated OpenAPI spec in CI, or inspecting captured traffic.

    ## Mental model

    A Mockxy workspace is a folder tree of files, and this API is a typed editor for it:

    - An **endpoint** is one HTTP method on one route path. On disk it is a
      `METHOD.endpoint.json` file; its route path is derived from the folder that contains it and
      is therefore **immutable after creation** (copy the endpoint to move it).
    - An endpoint owns one or more **response variants** (`NNN.response.json` inside
      `METHOD.responses/`). Exactly one is *selected* and answers requests. A `sequence`
      variant walks other variants request after request; selecting it is the only way to
      activate that scenario.
    - A variant has a `type` that decides how the endpoint answers:
      `mock` (static payload, optionally templated or file-backed), `handler` (JS module that
      computes the response), `middleware` (JS module that transforms a proxied backend
...
      `sequence` (an ordered scenario of `mock`/`handler` variants).
...

...

...

...

...

...
servers:
...
paths:
...
  /mocks:
    get:
      tags: [Catalog]
      operationId: listMocks
      summary: The whole catalog
      description: |
        Returns every endpoint with a summary of its **selected** variant, the collection tree,
        the unified child ordering used to render it, and the endpoints that could not be read.

        An unreadable endpoint file (invalid JSON, missing selected variant) never fails the
        request: it is skipped and reported in `loadErrors`, exactly as the runtime loader does.
      responses:
        "200":
          description: The catalog
          content:
            application/json:
              schema:
                type: object
                required: [items, collections, childOrder, loadErrors]
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/MockSummary" }
                    description: Endpoints, ordered as the UI shows them.
                  collections:
                    type: array
                    items: { $ref: "#/components/schemas/CollectionSummary" }
                    description: Persisted collections in depth-first display order.
                  childOrder: { $ref: "#/components/schemas/ChildOrder" }
                  loadErrors:
                    type: array
                    description: Endpoint files that could not be read, and why.
                    items:
                      type: object
                      required: [configFilePath, message]
                      properties:
                        configFilePath:
                          type: string
                          description: Path of the endpoint file, relative to the mocks folder.
                        message: { type: string }
        "404": { $ref: "#/components/responses/AdminApiDisabled" }
    post:
      tags: [Endpoints]
      operationId: createMock
      summary: Create an endpoint
      description: |
        Creates an endpoint together with its first variant. The route path determines the
        folder the files land in, so **method + path must be free**: an existing endpoint answers
        `409` and reports its id in `details.existingMockId`, which lets a client offer "add this
        as a variant of the existing endpoint" instead.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateMockRequest" }
            examples:
              staticMock:
                summary: Static JSON mock
                value:
                  config:
                    method: GET
                    path: /api/users
                    status: 200
                    headers: { content-type: application/json }
                    delayMs: 0
                  body:
                    - { id: 1, name: Ada }
                    - { id: 2, name: Linus }
              handler:
                summary: Dynamic handler (source omitted → template scaffold)
                value:
                  type: handler
                  definition: { method: POST, path: /api/login }
      responses:
        "201":
          description: The created endpoint
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MockDetail" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "409":
          description: An endpoint already exists for this method and path.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
              example:
                error: Conflict
                message: An endpoint already exists for this folder and method.
                details: { existingMockId: bW9ja3MvYXBpL3VzZXJzL0dFVC5lbmRwb2ludC5qc29u }
  /mocks/resolve:
    get:
      tags: [Catalog]
      operationId: resolveMock
      summary: Which endpoint would cover this request
      description: |
        Answers "if this concrete request arrived now, which catalog entry would it belong to?".
        The matching mirrors the serving semantics (exact routes before dynamic ones,
        specificity, declared query parameters) but runs over the **whole** catalog, disabled
        endpoints included — the use case is "take me to the mock", not "who would answer".

        A derived fact, computed on the fly and never persisted: the monitor entry stays the
        pure historical record.
      parameters:
        - name: method
          in: query
          required: true
          schema: { $ref: "#/components/schemas/HttpMethod" }
        - name: path
          in: query
          required: true
          description: Absolute request path; a query string may be included and is used by routes that declare query parameters.
          schema: { type: string }
          example: /api/users/42?expand=true
      responses:
        "200":
          description: The covering endpoint, or `null` when none matches.
          content:
            application/json:
              schema:
                type: object
                required: [mock]
                properties:
                  mock:
                    oneOf:
                      - { $ref: "#/components/schemas/MockSummary" }
                      - { type: "null" }
        "400": { $ref: "#/components/responses/BadRequest" }

  # ─────────────────────────────── Collections ───────────────────────────────
  /mocks/collections:
    post:
      tags: [Collections]
      operationId: createCollection
      summary: Create a collection
      description: |
        Creates a collection, optionally nested under another one. Labels are compared
        case- and accent-insensitively among siblings, and the reserved label of the virtual
        `Unsorted` bucket is rejected. New sub-collections land at the top of their parent;
        new top-level collections are appended.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [label]
              properties:
                label: { type: string }
                parentId:
                  type: string
                  description: Parent collection id; omit for a top-level collection. `unsorted` is not a valid parent.
            example: { label: Authentication }
      responses:
        "201":
          description: The created collection
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CollectionSummary" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404":
          description: Parent collection not found.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "409":
          description: The label is reserved, or already used by a sibling.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /mocks/collections/order:
    patch:
      tags: [Collections]
      operationId: reorderCollections
      summary: Reorder sibling collections
      description: Sets the order of the collections that share one parent. The list must contain each sibling **exactly once**.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [collectionIds]
              properties:
                collectionIds:
                  type: array
                  items: { type: string }
                  description: Persisted collection ids, in the desired order.
                parentId:
                  type: string
                  description: Scope of the reordering; omit for the top level.
      responses:
        "200": { $ref: "#/components/responses/CollectionList" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/CollectionNotFound" }
  /mocks/collections/{id}/parent:
    parameters:
      - $ref: "#/components/parameters/CollectionId"
    patch:
      tags: [Collections]
      operationId: reparentCollection
      summary: Move a collection in the tree
      description: |
        Re-parents a collection (omit `parentId` to move it to the top level) and optionally
        places it at a given index among its new siblings. Cycles are rejected: a collection
        cannot become a child of itself or of one of its descendants.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                parentId:
                  type: [string, "null"]
                  description: New parent; `null`/absent means top level. `unsorted` is not a valid parent.
                targetIndex:
                  type: integer
                  minimum: 0
                  description: Position among the new siblings; appended when absent.
      responses:
        "200": { $ref: "#/components/responses/CollectionList" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/CollectionNotFound" }
        "409":
          description: A sibling already uses this label.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /mocks/collections/{id}/items/order:
    parameters:
      - $ref: "#/components/parameters/CollectionIdWithUnsorted"
    patch:
      tags: [Collections]
      operationId: reorderCollectionItems
      summary: Reorder the endpoints of a collection
      description: Sets the order of the endpoints inside one collection. The list must contain each endpoint of that collection **exactly once**.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [itemIds]
              properties:
                itemIds:
                  type: array
                  items: { type: string }
                  description: Endpoint ids, in the desired order.
      responses:
        "200": { $ref: "#/components/responses/MockItemList" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/CollectionNotFound" }
  /mocks/collections/{parentKey}/children/order:
    parameters:
      - name: parentKey
        in: path
        required: true
        description: "`root`, `unsorted`, or a collection id."
        schema: { type: string }
    patch:
      tags: [Collections]
      operationId: reorderCollectionChildren
      summary: Reorder the interleaved children of a parent
      description: |
        Sets the unified order of everything directly under one parent — endpoints **and**
        sub-collections, interleaved as the tree shows them. `childRefs` mixes endpoint ids and
        collection ids and must contain each child of that parent exactly once.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [childRefs]
              properties:
                childRefs:
                  type: array
                  items: { type: string }
                  description: Endpoint ids and/or collection ids, in the desired order.
      responses:
        "200": { $ref: "#/components/responses/MockItemList" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/CollectionNotFound" }
  /mocks/collections/{id}/enabled:
    parameters:
      - $ref: "#/components/parameters/CollectionId"
    patch:
      tags: [Collections]
      operationId: setCollectionEnabled
      summary: Enable or disable a whole subtree
      description: |
        Rewrites the `enabled` flag of every endpoint in the collection **and its
        descendants** — the one collection operation that changes how requests are served.
        Endpoints already in the requested state are left untouched. The whole batch shares a
        single rollback and a single runtime reload.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [enabled]
              properties:
                enabled: { type: boolean }
      responses:
        "200":
          description: The refreshed catalog
          content:
            application/json:
              schema:
                type: object
                required: [items, collections, childOrder]
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/MockSummary" }
                  collections:
                    type: array
                    items: { $ref: "#/components/schemas/CollectionSummary" }
                  childOrder: { $ref: "#/components/schemas/ChildOrder" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/CollectionNotFound" }
  /mocks/collections/{id}:
    parameters:
      - $ref: "#/components/parameters/CollectionId"
    delete:
      tags: [Collections]
      operationId: deleteCollection
      summary: Dissolve a collection subtree
      description: |
        Removes the collection and its descendants **without deleting endpoints**: every
        endpoint they contained moves back to the virtual `Unsorted` bucket. To delete the
        endpoints too, use `DELETE /mocks/collections/{id}/contents`.
      responses:
        "204": { description: Dissolved. }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/CollectionNotFound" }
  /mocks/collections/{id}/contents:
    parameters:
      - $ref: "#/components/parameters/CollectionIdWithUnsorted"
    delete:
      tags: [Collections]
      operationId: eraseCollection
      summary: Delete a subtree and every endpoint in it
      description: |
        Permanently deletes the collection subtree **and all the endpoints it contains**, files
        included. With `id=unsorted` it deletes exactly the endpoints that belong to no
        collection. One shared rollback, one runtime reload.
      responses:
        "200":
          description: How many endpoints were deleted.
          content:
            application/json:
              schema:
                type: object
                required: [deleted]
                properties:
                  deleted: { type: integer }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/CollectionNotFound" }

  # ──────────────────────────────── Endpoints ────────────────────────────────
...
  /mocks/{id}/sequence/state:
    parameters:
      - $ref: "#/components/parameters/MockId"
    get:
      tags: [Endpoints]
      operationId: getSequenceState
      summary: Read the live cursor of the selected sequence
      responses:
        "200":
          description: Selected sequence file and live cursor
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SequenceStateResponse" }
        "400":
          description: The selected response is not a sequence.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404": { $ref: "#/components/responses/MockNotFound" }
...
components:
...
  responses:
    BadRequest:
      description: Invalid input.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
...
    HttpMethod:
      type: string
      enum: [GET, POST, PUT, PATCH, DELETE, HEAD, OPTIONS]
...
    CollectionSummary:
      type: object
      required: [id, label, itemCount]
      properties:
        id: { type: string }
        label: { type: string }
        itemCount:
          type: integer
          description: Endpoints assigned **directly** to this collection.
        parentId:
          type: string
          description: Present only for nested collections.
...